requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
backoff==2.2.1
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==2.2.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.2
dagshub-annotation-converter==0.1.11
databricks-cli==0.18.0
databricks-sdk==0.60.0
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
Flask==2.3.3
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.30.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lxml==6.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.3.2
mlflow-skinny==3.1.4
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-nccl-cu12==2.27.6
oauthlib==3.3.1
opentelemetry-api==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==23.2
pandas==2.3.1
parso==0.8.4
pathvalidate==3.3.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.4
PyYAML==6.0.2
pyzmq==27.0.0
querystring-parser==1.2.4
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.0.2
yarl==1.20.1
zipp==3.23.0
